removing redundant package lists and sentence-transformers library (#8971)
* removing redundant package lists and sentence-transformers library

* Add input column to RAG tests and update requirements

Added 'input_column' to RAG tests in `test_rag.py` to accommodate for changes in the function signatures. Furthermore, the RAG handler requirements have been updated to include 'sentence-transformers', required for HuggingFaceEmbeddings from the langchain-community package.

* Add allow_dangerous_deserialization in FAISS loading and fix typo in unit tests.

A new allow_dangerous_deserialization parameter was added to the FAISS loading in RAG handler settings, providing additional flexibility for deserialization. Additionally, a typo in the input column name within unit tests was fixed, which was causing the tests to fail.

---------

Co-authored-by: dusvyat <13661123+dusvyat@users.noreply.github.com>